+++ packages/ethereum-contracts/test/contracts/superfluid/Superfluid.test.ts
@@ -112,11 +112,13 @@ describe("Superfluid Host Contract", function () {
                     await ethers.getContractFactory("SuperfluidMock");
                 const mock1 = await sfMockFactory.deploy(
                     false /* nonUpgradable */,
-                    false /* appWhiteListingEnabled */
+                    false /* appWhiteListingEnabled */,
+                    ZERO_ADDRESS /* dmzForwader */
                 );
                 const mock2 = await sfMockFactory.deploy(
                     true /* nonUpgradable */,
-                    false /* appWhiteListingEnabled */
+                    false /* appWhiteListingEnabled */,
+                    ZERO_ADDRESS /* dmzForwader */
                 );
                 await governance.updateContracts(
                     superfluid.address,
@@ -662,6 +664,7 @@ describe("Superfluid Host Contract", function () {
             });
         });
 
+        // disabled due to contract size limit
         describe("#5 Context Utilities", () => {
             it("#5.1 test replacePlaceholderCtx with testCtxFuncX", async () => {
                 const testCtxFunc = async (
@@ -694,6 +697,7 @@ describe("Superfluid Host Contract", function () {
                     );
                 }
 
+                // disabled code because contract size limit hit
                 // more complicated ABI
                 await testCtxFunc(
                     "ctxFunc2",
@@ -701,8 +705,8 @@ describe("Superfluid Host Contract", function () {
                         governance.address,
                         t.contracts.ida.address,
                         ethers.utils.hexZeroPad("0x2020", 32),
-                        "0x" /* agreementData */,
-                        "0x" /* cbdata */,
+                        "0x", // agreementData
+                        "0x", // cbdata
                     ],
                     "0x" + "dead".repeat(20)
                 );
@@ -712,8 +716,8 @@ describe("Superfluid Host Contract", function () {
                         governance.address,
                         t.contracts.ida.address,
                         ethers.utils.hexZeroPad("0x2020", 32),
-                        "0xdead" /* agreementData */,
-                        "0xbeef" /* cbdata */,
+                        "0xdead", // agreementData
+                        "0xbeef", // cbdata
                     ],
                     "0x" + "faec".repeat(20)
                 );
@@ -2697,7 +2701,8 @@ describe("Superfluid Host Contract", function () {
                     await ethers.getContractFactory("SuperfluidMock");
                 const mock1 = await mock1Factory.deploy(
                     false /* nonUpgradable */,
-                    false /* appWhiteListingEnabled */
+                    false /* appWhiteListingEnabled */,
+                    ZERO_ADDRESS /* dmzForwader */
                 );
                 await expectCustomError(
                     governance.updateContracts(